-Wall -Wextra -Werror

+++ includes/Request.hpp
@@ -31,7 +31,7 @@ class Request
 		bool _chunk_finish_flag;
 		size_t _chunk_size;
 		size_t _status_number;
-		int _body_size;
+		size_t _body_size;
 		bool _connection_flag;
 		bool _bad_request_flag;
 		std::vector<std::string> _valid_header_key;
@@ -114,8 +114,8 @@ void setValidHeaderKey(const std::vector<std::string>& valid_header_key);
 bool getBadRequestFlag(void) const;
 void setBadRequestFlag(const bool& bad_request_flag);
 
-int getBodySize(void) const;
-void setBodySize(const int& body_size);
+size_t getBodySize(void) const;
+void setBodySize(const size_t& body_size);
